requirements.txt
spade==3.2.3
spade_pubsub==0.1.3
spade_artifact==0.1.2
pandas==2.0.0
prophet==1.1.2
python-dateutil==2.8.2
